Updated site object to accomodate is_inverter option + created /inverters/enphase.py file
